Provide -onbuild variants of all images
This lets people use our images as a base, but
customize it easily with a subset of the files
understood by repo2docker. Currently, this includes:

- environment.yml
- requirements.txt
- postBuild

We can add more if necessary.

The contents of the user's repository are also copied
into ${HOME}, merging with (and overwriting) anything
that's in our base images.

To use the onbuild images as a base,

1. In your git repository, create a file named 'Dockerfile'
2. Write the name of the base image you would like to use
   in the following format:

   'FROM <base-image>:<tag>'

   For example, you can write:

   'FROM pangeo/base-notebook-onbuild:2019.04.15'

   Note the -onbuild - that is very important.
3. Create a 'environment.yml', 'requirements.txt' or
   'postBuild' file in your repository (or under a
   'binder/' directory) to customize your environment
   when launched in a binder.

That's it! When this repository is launched in any
binder, your customizations will be applied on top
of the base image you have chosen, and launched.

Fixes #26